Add a libc++ builder for Linux.
The existing libc++ builder is based on buildit/testit and assumes that
the platform has its own libc++abi, and as such it only works for OS X.
Add a new one that uses an in-tree build and cmake to be compatible with
the other platforms.

This builder will actually build and test both libc++ and libc++abi.
Since most platforms don't have either of these libraries installed and
depend on each other, we can't test them in isolation.